requirements.txt
tensorflow==2.2.0
Flask
Flask-JSON
pillow
